修改了一个BUG:根据分类id查询菜品时传入的CategoryID类型为Integer,应为Long 添加了套餐删除业务 修改了分类删除业务的逻辑,现在当分类被菜品或套餐关联时,分类无法删除

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -37,7 +37,7 @@ public class DishController {
 
     @GetMapping("/list")
     @ApiOperation("根据分类id查询菜品")
-    public Result<List> list(Integer categoryId){
+    public Result<List> list(Long categoryId){
         log.info("根据分类ID查询,参数为{}",categoryId);
         List<Dish> list=dishService.list(categoryId);
         return Result.success(list);
